refactor run-tests api
